+++ src/features/user/ui/EditProfile.styles.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const NicknameSection = styled.div`
+  margin-top: 10px;
+  display: flex;
+  gap: 10px;
+  align-items: start;
+`;
+
+export const ButtonWrapper = styled.div`
+  width: 100px;
+`;
